1. StickerRepository 생성 2. StickerDto에 totalCount, successCount 삭제, stickerColor 추가 3. 1시간, 자정마다 데이터베이스 자동 생성되는 로직 CronComponent로 이동
